+++ app/src/main/AndroidManifest.xml
@@ -19,21 +19,12 @@
         <service android:name=".WeatherLoaderService"></service>
 
         <activity
-            android:name="ru.ifmo.md.lesson8.ItemListActivity"
+            android:name=".MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity
-            android:name="ru.ifmo.md.lesson8.ItemDetailActivity"
-            android:label="@string/title_item_detail"
-            android:parentActivityName="ru.ifmo.md.lesson8.ItemListActivity">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ItemListActivity" />
         </activity>
 
     </application>
