+++ app/src/main/java/ru/ifmo/md/lesson8/logic/CityWeather.java
@@ -1,4 +1,4 @@
-package ru.ifmo.md.lesson8;
+package ru.ifmo.md.lesson8.logic;
 
 import java.util.ArrayList;
