+++ app/src/main/res/values/strings.xml
@@ -2,7 +2,6 @@
 <resources>
 
     <string name="app_name">Weather</string>
-    <string name="title_item_detail">Item Detail</string>
     <string name="action_settings">Settings</string>
     <string-array name="views_array">
         <item>Первый экран</item>
@@ -11,7 +10,7 @@
     </string-array>
     <string name="toggle_open">Open city list</string>
     <string name="toggle_close">Close city list</string>
-    <string name="menu">Меню</string>
+    <string name="city_list">Cities</string>
     <string name="description_refresh">lalala</string>
 
 </resources>
